refactor: Split tools/utils.ts into naming and input-schema modules
What: src/tools/utils.ts mixed two unrelated concerns behind a vague name —
a few tool-name/Actor-predicate helpers and the bulk Actor-input-schema
transformation pipeline — so consumers imported the same module for very
different reasons.

How: split into src/tools/actor_tool_naming.ts (actorNameToToolName,
legacyToolNameToNew, getToolSchemaID, isActorInfoMcpServer,
isActorBlockedUnderPaymentProvider) and src/tools/actor_input_schema.ts (the
transform/AJV/enum bulk, kept as the git-mv rename). One-way dependency:
the schema module imports getToolSchemaID from the naming module. Importers
re-pointed; the internals re-export surface is unchanged, so the hosted
server needs no change. Pure structural move — behavior identical.

Alternatives considered: a re-export barrel at the old path (rejected — no
external subpath can deep-import it, so it would be dead indirection); a third
module for the two Actor predicates (rejected — two one-liners with no schema
coupling don't warrant it).

Closes #1021
