+++ src/index_internals.ts
@@ -8,6 +8,7 @@ import { processParamsGetTools } from './mcp/utils.js';
 import { resolvePaymentProvider } from './payments/index.js';
 import type { PaymentProvider } from './payments/types.js';
 import { getServerCard } from './server_card.js';
+import { actorNameToToolName } from './tools/actor_tool_naming.js';
 import { addActor } from './tools/actors/add_actor.js';
 import {
     getActorsAsTools,
@@ -17,7 +18,6 @@ import {
     toolCategoriesEnabledByDefault,
     unauthEnabledTools,
 } from './tools/index.js';
-import { actorNameToToolName } from './tools/utils.js';
 import type { ActorStore, ServerCard, ToolCategory } from './types.js';
 import { parseCommaSeparatedList, parseQueryParamList, readJsonFile } from './utils/generic.js';
 import { redactSkyfirePayId } from './utils/logging.js';
